init: Tune internal oscillator for new PICs
When transistioning to breadboard development
a brand new PIC16F690 was used.  In order for
the UART to correctly function, the internal
oscillator needs to be tuned down to its
minimum frequency.  However, the UART on the
original test board PIC functions correctly
at the factory-calibrated default tuning level.

Create preprocessor wrappings to deal with
this difference.  The cause of this may be
investigated later, however the final product
will use the new chips.