build: Add rule for building test firmware
The regular firmware can be built using 'make dexterity' or simply
'make'.  This will compile the firmware using dexterity.c as the main
program.  To build the test firmware using test.c as the main program,
run 'make test TYPE=<test_type>', where 'test_type' is the defined macro
in test.c that specifies which test function to run.

This eliminates the need to edit the source code of test.c in order to
switch the test code.