common: Use standard size for transmitted data fields
On the PIC an int is a 16b number, while on the host it
is usually a 32b number.  These sizes must be consistent
in order to avoid data misinterpretation.

Use the int16_t type to represent all integer data that
will be transmitted and received over UART between the
device and the host.