firmware: Add functionality for applying a calibration payload
The PIC will first wait for the appropriate message and send an
ACK message to notify the master that it is ready for the data.
It will then wait to receive the data and will apply it once
the entire packet arrives.

TODO: We should add ACKs for all the functions of the device.
      This will be useful for debugging instead of having a
      'fire-and-forget' mentality.