common: Put enum lengths inside enumeration list
The number of enumerations for each type of enum were originally defined
as macros.  This may have been causing warnings, especially if enum
numbering is different between gcc and xc8.  By making the last enum
member the length of the enum, we can always ensure that the correct
length is used across the firmware and the software.