software: Purge serial port before every write
serial_purge() was inserted directly into serial_write().
This avoids the need to manually call it before every write.
The logic behind purging is to clear an pending data that may be
stuck in the TX buffer.  This ensures that the new data is transmitted
as expected without any additional remnants from previous transfers.

The RX buffer is never purged before data is read from the serial port,
as read() may block, so we don't want to lose data if the entire
transmission is still coming in.  write() will simply dump the data in
the TX buffer and return, so we don't know for sure when it has been
transmitted.  This is the reason we use serial_purge().