software: Switch between control modes using thumb
There will be 4 control modes:
MOUSE - move the cursor and click
SCROLL - scroll the mouse wheel
ZOOM - zoom in and out
VOLUME - control the volume

It is easier to cycle between the 4 modes instead of
attempting to transition based on sensor data.  This
will also allow fewer profiles to be checked at a time,
which should decrease latency.

These modes can be switched between using a thumb flick,
but this is configurable via the gesture file.