software: Create help and usage messages
These can be triggered using the '--help' command line argument.
Each subcommand has its own help message which shows the user how
to use that subcommand and what it does.  Using '--help' on the
root command will bring up a menu of all the supported commands.

NOTE: the 'mode' subcommand was reverted back to the 'raw' and
      'scaled' subcommands for implementation simplicity.